move logic files to lib
